+++ crates/ruma-events/CHANGELOG.md
@@ -3,6 +3,8 @@
 Improvements:
 
  - Add support for encrypted stickers as sent by several bridges under the flag `compat-encrypted-stickers`
+ - Add unstable support for MSC3489 `m.beacon` & `m.beacon_info` events
+   (unstable types `org.matrix.msc3489.beacon` & `org.matrix.msc3489.beacon_info`)
 
 Breaking changes: